requirements.txt
openai~=1.2.3
python-dotenv~=1.0.0
distro~=1.8.0
colorama~=0.4.6
idna~=3.4
certifi~=2023.7.22
tiktoken~=0.5.1
PyYAML~=6.0.1
Jinja2~=3.1.2
requests~=2.31.0
regex~=2023.10.3
click~=8.1.7
urllib3~=2.0.7
flask~=3.0.0
celery~=5.3.5